Add test command to tinygo
* Test command passes flag to compile indicating that a test binary
should be built
* Adds the test build tag
* Imports packages from test files

Stuck on how to find the TestMain that I have defined in my test file.
It is coming back nil.

Not sure how I would reflect on the functions defined in the test files
and then build up the test suite (M) in a basic block. First step I
figured was calling a simple predefined test suite though.